[jinterface] Kill process between tests to avoid consequential error
In some cases when a test case fails, the erl_link_server process
(which is spawned by many test cases in jinterface_SUITE) does not
terminate. This causes the next test case to fail with a badarg as it
tries to register a new process with the same name. To avoid this,
erl_link_server, if it exists, is now killed in end_per_testcase.

Also, some compiler warnings are removed from jitu.erl.